+++ src/api/messages/markAsRead.js
@@ -1,6 +1,6 @@
 /* @flow */
-import type { Auth } from '../types';
-import messagesFlags from './messages/messagesFlags';
+import type { Auth } from '../../types';
+import messagesFlags from './messagesFlags';
 
 export default (auth: Auth, messageIds: number[]): any =>
   messagesFlags(auth, messageIds, 'add', 'read');
