+++ src/message/messagesActions.js
@@ -8,6 +8,10 @@ import openLink from '../utils/openLink';
 import { fetchMessagesAtFirstUnread } from './fetchActions';
 import { validateNarrow } from '../utils/narrow';
 // import { showToast } from '../utils/info';
+import { markAsRead } from '../api';
+
+let unsentMessageIds = [];
+let isMarkAsReadLooping = false;
 
 export const switchNarrow = (narrow: Narrow): Action => ({
   type: SWITCH_NARROW,
@@ -50,16 +54,37 @@ export const messageLinkPress = (href: string) => (dispatch: Dispatch, getState:
   }
 };
 
-export const addReadFlagToMessages = (narrow: Narrow, messageIds: number[]): Action => ({
+export const addReadFlagToMessages = (messageIds: number[]): Action => ({
   type: EVENT_UPDATE_MESSAGE_FLAGS,
   messages: messageIds,
   flag: 'read',
   operation: 'add',
 });
 
-export const markMessageAsRead = (narrow: Narrow, messageIds: number[]): Action => (
+export const markMessageAsReadQueue = () => async (dispatch: Dispatch, getState: GetState) => {
+  // loop to mark message as read
+  // eslint-disable-next-line no-constant-condition
+  while (true) {
+    dispatch(addReadFlagToMessages(unsentMessageIds));
+    markAsRead(getAuth(getState()), unsentMessageIds);
+    unsentMessageIds = [];
+    // eslint-disable-next-line no-await-in-loop
+    await new Promise(resolve => setTimeout(resolve, 2000));
+    if (unsentMessageIds.length === 0) {
+      isMarkAsReadLooping = false;
+      break;
+    }
+  }
+};
+
+export const addMessagesToQueue = (messageIds: number[]): Action => (
   dispatch: Dispatch,
   getState: GetState,
 ) => {
-  dispatch(addReadFlagToMessages(narrow, messageIds));
+  unsentMessageIds.push(...messageIds);
+  if (!isMarkAsReadLooping) {
+    // start loop
+    dispatch(markMessageAsReadQueue());
+    isMarkAsReadLooping = true;
+  }
 };
